Further reorg for modularizing logic
